requirements.txt
langchain
langchain-community
langchain-ollama
faiss-cpu
tiktoken
watchdog
pydantic
fastapi
uvicorn